Fix the comments on logical connectives
This small patch fixes the misleading comments on the logical
connectives. Currently, they're both used in SMT-lib as bitwise and as
logical ones depending on the return type. It's likely to be a hack.
Though the comment says they're used as bitwise functions only.
Let's change it to the more descriptive not to be misleading.